{plat, lib}: `__u64` for exceptions `error_code` struct member
Since we exclusively only support 64-bit platforms/architectures and
everything else is written around it, be more exact and consistent and
also apply this to the `error_code` struct members of exception structs.

Lastly, also remove the exception handler prototype from the Xen
platform's traps.h as it does not seem to be needed and we will likely
remove this header entirely anyway as well as the unnecessarily
duplicated error_code getter/setters that are not x86_64 namespaced.

+++ lib/ukpal/arch/x86_64/include/uk/pal/arch/except.h
@@ -28,12 +28,12 @@ __isr void uk_pal_x86_64_except_err_ctx_set_trapnr(
 	struct uk_pal_except_err_ctx *ctx,
 	__u32 trapnr);
 
-__isr int uk_pal_except_err_ctx_get_error_code(
+__isr __u64 uk_pal_except_err_ctx_get_error_code(
 	const struct uk_pal_except_err_ctx *ctx);
 
 __isr void uk_pal_except_err_ctx_set_error_code(
 	struct uk_pal_except_err_ctx *ctx,
-	int error_code);
+	__u64 error_code);
 
 #ifdef __cplusplus
 }
